# Previous point and quick cache