--- a/lib/Solver/CexCachingSolver.cpp
+++ b/lib/Solver/CexCachingSolver.cpp
@@ -144,6 +144,9 @@ class CexCachingSolver : public SolverImpl {
 
   bool quickMatch(const Query &query, const KeyType &key, Assignment *&result);
 
+  bool checkPreviousSolutionHelper(const ref<Expr>, const std::set<ref<Expr> > &key, Assignment * &result);
+  bool checkPreviousSolution(const Query &query, Assignment *&result);
+
   bool getAssignment(const Query& query, Assignment *&result);
   
 public:
@@ -293,6 +296,84 @@ bool CexCachingSolver::quickMatch(const Query &query,
 	return false;
 }
 
+bool CexCachingSolver::checkPreviousSolutionHelper(const ref<Expr> queryExpr,	//If anything other than query.expr, then negated.
+										   const std::set<ref<Expr> > &key,
+										   Assignment * &result){
+	Assignment * parentSolution;
+	if(getFromQuickCache(key, parentSolution)){
+		if(!parentSolution){
+			//means that the the previous state was UNSAT and therefore the
+			//new answer will also necessarily be UNSAT
+			assert(!result);
+			return true;
+		}else{
+			/*
+			 * Means that there is in fact a parent solution.  In this case,
+			 * we can now check whether this parent solution satisfies the
+			 * child state.  There's a pretty good chance... at least 50/50
+			 */
+			ref<Expr> neg = Expr::createIsZero(queryExpr);
+			ref<Expr> q = parentSolution->evaluate(neg);
+
+			assert(isa<ConstantExpr>(q) && "assignment evaluation did not result in constant");
+			if(cast<ConstantExpr>(q)->isTrue()){
+				result = parentSolution;
+				return true;
+			}else{
+				/*
+				 * If goes false, that means that the point that had gotten us to our parent
+				 * went along the opposing branch and it won't help us at this stage.  The
+				 * value stored in oldAnswer could help someone though.
+				 */
+				assert(!result);
+				return false;
+			}
+		}
+	}
+	assert(!parentSolution);
+	assert(!result);
+	return false;
+}
+
+bool CexCachingSolver::checkPreviousSolution(const Query &query,
+										  Assignment *&result){
+	if(query.constraints.size() == 0){
+		return false;
+	}
+
+	std::set<ref<Expr> > parentKey;
+	ref<Expr> queryExpr;
+	if(isa<ConstantExpr>(query.expr)){
+		assert(cast<ConstantExpr>(query.expr)->isFalse() && "query.expr == true should'd happen");
+		for(unsigned i = 0; i < query.constraints.size() - 1; i ++){
+			ref<Expr> ref = query.constraints.get(i);
+			parentKey.insert(ref);
+		}
+
+        ref<Expr> toNeg = query.constraints.get(query.constraints.size() - 1);
+        queryExpr = Expr::createIsZero(toNeg);
+	}else{
+		for(unsigned i = 0; i < query.constraints.size(); i ++){
+			ref<Expr> ref = query.constraints.get(i);
+			parentKey.insert(ref);
+		}
+		queryExpr = query.expr;
+	}
+
+	if(checkPreviousSolutionHelper(queryExpr, parentKey, result)){
+		/*
+		 * result may contain one of two things
+		 * 	- A 0, meaning that the previous piece of the was UNSAT and therefore new is too
+		 *	  (I put an assertion in checkPrevHelper to discount this case, but being careful)
+		 * 	- An actual result which we should verify is correct.
+		 */
+		return true;
+	}else{
+		return false;
+	}
+}
+
+
 /// lookupAssignment - Lookup a cached result for the given \arg query.
 ///
 /// \param query - The query to lookup.
@@ -317,6 +398,13 @@ bool CexCachingSolver::lookupAssignment(const Query &query,
   }
 
   bool found = quickMatch(query, key, result);
+
+  if(!found){
+  		found = checkPreviousSolution(query, result);
+  		if(found){
+  			insertInQuickCache(key, result);
+  		}
+  }
 
   if(!found){
 	  found = searchForAssignment(key, result);