This commit is conceptually orthogonal to the other pull request submitted on 2/16/15 called IndependentSolverGetIntialValues.  In actuality, however, they play off of each other for a combined effect superior to either one individually and involve a lot of the same files.

This particular commit targets the CexCachingSolver so that the UBTree is called less frequently. Specifically, the two optimizations are a "QuickCache" that determines when two constraints are identical. The cache stores solutions so that they can be used in the future. The second optimizations tests the parent state's solution on every newly created state. This will work, by defn, at least
50% of the time. It also prevents the UBTree's invocation, which can be costly.

When the improvements present in this in pull request are combined with the IndependentSolverGetInitalValues pull request, the overall time spent in the default search is on average, almost halved.

After completion, the changes:
1) Pass all provided tests (other than ones that were expected to fail)
2) Have been verified via -debug-validate-solver by a complete run through
of 9 members of the bin-utils for
--search=dfs with a halt being placed, and states dumped, after 500 forks
--search=default with a halt being placed, and the states dumped, after 1000 forks
3) Achieved significant speedup on test programs
--search=dfs
a. On average 93% improvement (for example, nm-new saw a decrease from 1200 seconds to 670 seconds)
b. Worst case 13% improvement (obj-copy when from 950 seconds to 835 seconds)
--search=default
a. On average 85% improvement (for example, addr2line saw a decrease from 1325 seconds to 717 seconds)
b. Worst case 29% improvement (ar went from 560 seconds to 435 seconds)