@ccadar 2015-02-12
Thank you for the patch, @holycrap872 .  This looks like an excellent improvement, and I'm looking forward to seeing it in the mainline.  I also appreciate the numbers provided, they are very useful!

However, I would prefer to break things down into smaller patches.  This would both speed up the review process (large patches are not that easy to review, and I also tend to postpone reviewing them :)), but even more importantly, they are easier to understand and also to debug/revert if we detect an error later on.