@holycrap872 2015-02-12
No trouble @ccadar .  I'm definitely a bit of a GitHub newb, so I appreciate you being patient with me.  I just pushed a much more detailed commit history that will hopefully be much easier to review.  Again, if you would like me to update/improve anything, just let me know.