@ccadar 2015-02-15
Thanks, @holycrap872.  If these commits depend on each other and should be applied in this order, then this is fine.  But if there are several independent (even though related) commits, I prefer to receive them as separate pull requests for a couple of reasons: (1) they can be reviewed and approved independently, and any discussions and improvements affect only the pull requests concerned and (2) Travis CI triggers a separate regression test run for each commit which is better from a testing perspective.