@holycrap872 2015-02-16
Hopefully this is a bit easier @ccadar .  Again, thanks for being understanding and sorry for all the hassle.