@ccadar 2015-03-02
Hi @holycrap872 , what I propose with this pull request is to resubmit it as two different requests, one for each optimisation. Does this make sense?