@holycrap872 2015-03-02
I can do that, but the Previous Point optimization is dependent on the QuickCache.  

This is because ALL satisfying points (even different sets of constraints with the same solution) are saved into the QuickCache which allows for quick access to the model that satisfied the previous state.  There isn't really a comparable implementation that uses the current UBTree setup.  The reason for this is that we would to keep the UBTree as small as possible.  The absence of a QuickCache would necessitate adding every satisfying point to the UBTree and then relying on it's current lookup function.  This would cause the UBTree to bloat with multiple sets of constraints pointing to the same solution.