@ccadar 2015-03-02
I see.  Let's start just with the QuickCache optimisation then.  Thanks!